日期

+++ src/main/java/com/hthyaq/zybadmin/controller/dataVisual/nationyes/NationGovVisualYes.java
@@ -97,7 +97,7 @@ public class NationGovVisualYes {
             govEight.setVar3(RandomUtil.randomInt(1, 10000));
             govEight.setVar4(RandomUtil.randomInt(1, 10000));
             govEight.setVar5(RandomUtil.randomInt(1, 10000));
-            govEight.setVar6((double) RandomUtil.randomInt(1, 10000));
+            govEight.setVar6(RandomUtil.randomInt(1, 10000));
             govEight.setVar7(RandomUtil.randomInt(1, 10000));
             govEight.setVar8(RandomUtil.randomInt(1, 10000));
             list.add(govEight);
